Renamed test/ into tests/
