+++ tests/data/rasters/test3.tif
