+++ tests/test_geo_utils.py
@@ -3,7 +3,7 @@ Test geo utils.
 """
 from qgis.core import QgsPointXY
 from qgis.core import QgsVectorLayer
-from ThreeDiToolbox.test.utilities import ensure_qgis_app_is_initialized
+from ThreeDiToolbox.tests.utilities import ensure_qgis_app_is_initialized
 from ThreeDiToolbox.utils.geo_utils import calculate_perpendicular_line
 from ThreeDiToolbox.utils.geo_utils import get_coord_transformation_instance
 from ThreeDiToolbox.utils.geo_utils import get_distance
