+++ tests/test_misc_tools.py
