+++ tool_commands/guess_indicators/test_guess_indicators.py
@@ -6,7 +6,7 @@ from ThreeDiToolbox.sql_models.model_schematisation import ConnectionNode
 from ThreeDiToolbox.sql_models.model_schematisation import Manhole
 from ThreeDiToolbox.sql_models.model_schematisation import Pipe
 from ThreeDiToolbox.sql_models.model_schematisation import Pumpstation
-from ThreeDiToolbox.test.test_init import TEST_DATA_DIR
+from ThreeDiToolbox.tests.test_init import TEST_DATA_DIR
 from ThreeDiToolbox.tool_commands.guess_indicators import guess_indicators_utils
 from ThreeDiToolbox.utils.threedi_database import ThreediDatabase
