+++ tool_result_selection/test_models.py
@@ -1,5 +1,5 @@
 from ThreeDiToolbox.datasource.threedi_results import ThreediResult
-from ThreeDiToolbox.test.test_datasources import THREEDI_RESULTS_PATH
+from ThreeDiToolbox.tests.test_datasources import THREEDI_RESULTS_PATH
 from ThreeDiToolbox.tool_result_selection.models import TimeseriesDatasourceModel
 
 import unittest
